requirements.txt
streamlit>=1.36
groq>=0.9.0
python-dotenv

